- fixed missing conversions of the uses of converted entities (of conversions of getters to properties, esp. of GetIRType) - missing adaptations from camelCase to PascalCase - specified the type argument for the GetResult<T> method explicitly - introduced explicit goto for a switch-case fall-through - adaptation to a RemoveFirst without return value - missing adaptation of the migration of File handling to .NET - type fix related to Java wildcard generics (type sharpening) - introduced an explicit cast when creating a boxed/nullable type - removed a nullable scalar stemming from the conversion of a Java wrapper class - changed access modifier to public - removed override

+++ frontend/de/unika/ipd/grgen/ast/expr/numeric/FloatConstNode.cs
@@ -19,7 +19,7 @@ namespace de.unika.ipd.grgen.ast.expr.numeric
 	public class FloatConstNode : ConstNode
 	{
 		public FloatConstNode(Coords coords, double v)
-			: base(coords, "float", new float?(v))
+			: base(coords, "float", new float?((float)v))
 		{
 		}
